Move error definition to a separated module out of binary.
