+++ src/converter.rs
@@ -2,3 +2,7 @@
 
 pub mod csv;
 pub mod iso_camt053;
+
+mod error;
+
+pub use error::ConvertError;
