+++ src/data.rs
@@ -1,23 +1,20 @@
-extern crate chrono;
-extern crate rust_decimal;
-
 use rust_decimal::Decimal;
 
 #[derive(Debug, PartialEq)]
 pub struct Transaction {
-    date: chrono::NaiveDate,
-    payee: String,
-    posts: Vec<Post>,
+    pub date: chrono::NaiveDate,
+    pub payee: String,
+    pub posts: Vec<Post>,
 }
 
 #[derive(Debug, PartialEq)]
 pub struct Post {
-    account: String,
-    amount: Amount,
+    pub account: String,
+    pub amount: Amount,
 }
 
 #[derive(Debug, PartialEq)]
 pub struct Amount {
-    value: Decimal,
-    commodity: String,
+    pub value: Decimal,
+    pub commodity: String,
 }
